# api-definitions/openapi.yaml
openapi: 3.0.3
info:
  title: sdmay23-40 Interactive Classroom Backend - OpenAPI 3.0
  description: |-
    These are the api documents for the sdmay23-40 backend
    Some useful links:
    - [sdmay23-40 repository](https://git.ece.iastate.edu/sd/sdmay23-40)
  contact:
    email: miller.tyler094@gmail.com
  version: 1.0.0
servers:
  - url: https://sdmay23-40.ece.iastate.edu/api
...
paths:
...
  /role:
    get:
      tags:
        - role
      summary: get role info
      description: get a roles info by their id
      operationId: getRole
      parameters:
        - in: query
          required: true
          name: role_id
          schema:
            type: integer
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  role_id:
                    type: integer
                    example: 2
                  course_id:
                    type: integer
                    example: 1
                  user_id:
                    type: integer
                    example: 4
                  role: 
                    type: string 
                    example: PROFESSOR | TA | STUDENT
        '400':
          description: Invalid Parameters
        '401':
          description: Unauthorized
        '403':
          description: Invalid Session
        '500':
          description: Internal Server Error
    post:
      tags:
        - role
      summary: Assigns the logged in user a new role to the system
      parameters:
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      requestBody:
        description: Create a new role in the system
        content:
          application/json:
            schema:
              properties:
                course_id:
                  type: integer
                  example: 2
                user_id:
                  type: integer
                  example: 2
                role:
                  type: string
                  example: STUDENT/TA/PROFESSOR
        required: true
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  rollID:
                    type: integer
                    example: 1
        '400':
          description: Invalid Body
        '401':
          description: Unauthorized
        '403':
          description: Invalid Session
        '500':
          description: Internal Server Error
...
  /message/messagesByLecture:
    get:
      tags:
        - message
      summary: gets all messages and polls for a lecture
      description: gets all messages and polls for a lecture
      operationId: getMessagesAndPollsByLectureId
      parameters:
        - in: query
          required: true
          name: lecture_id
          schema:
            type: string
        - in: query
          required: false
          name: timestamp
          schema:
            type: string
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                  type: array
                  items:
                    anyOf:
                      - $ref: '#/components/schemas/Message'
                      - $ref: '#/components/schemas/Poll'
        '400':
          description: Invalid Parameters
        '403':
          description: Invalid Session
        '500':
          description: Internal Server Error
  /download-media:
    get:
      tags:
        - media
      summary: download media
      description: download media by media_id & course_id
      operationId: downloadMedia
      parameters:
        - in: query
          required: true
          name: course_id
          schema:
            type: integer
        - in: query
          required: true
          name: media_id
          schema:
            type: integer
        - in: header
          name: x-session-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: string
                format: binary
...
components:
  schemas:
    Message:
      type: object
      properties:
        sender_id:
          type: integer
          example: 2
        timestamp:
          type: string
          example: 2017-07-23 13:10:11
        role: 
          type: string
          example: owner
        netId:
          type: string
          example: twmiller
        body: 
          type: string
          example: This is an example message
        is_anonymous:
          type: boolean
          example: false
        parent_id:
          type: integer
          example: 3 | null
        message_id:
          type: integer
          example: 2
        media_id:
          type: string
          example: string | null
        file_type:
          type: string
          example: string | null
    Poll:
      type: object
      properties:
        poll_id:
          type: integer
          example: 2
        timestamp:
          type: string
          example: 2017-07-23 13:10:11
        question:
          type: string
          example: what is the speed of gravity?
        close_date:
          type: string
          nullable: true
        poll_type:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              poll_choice_id:
                type: integer
                example: 1
              text:
                type: string
                example: 9.82m/s^2
              is_correct_choice:
                type: boolean
                example: true
...